# [Refactor] refactor planner profile