--- a/fe/fe-core/src/main/java/com/starrocks/common/profile/Timer.java
+++ b/fe/fe-core/src/main/java/com/starrocks/common/profile/Timer.java
@@ -1,0 +1,45 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package com.starrocks.common.profile;
+
+import java.util.Collections;
+import java.util.List;
+
+public class Timer implements AutoCloseable {
+    public String name() {
+        return "";
+    }
+
+    public void start() {
+    }
+
+    public void close() {
+    }
+
+    public long getTotalTime() {
+        return 0;
+    }
+
+    public long getFirstTimePoint() {
+        return 0;
+    }
+
+    public List<Long> getAllTimePoint() {
+        return Collections.emptyList();
+    }
+}