… fe/fe-core/src/main/java/com/starrocks/common/profile/VarTracer.java
… fe/fe-core/src/main/java/com/starrocks/connector/RemoteFileOperations.java
… fe/fe-core/src/main/java/com/starrocks/connector/iceberg/IcebergMetadata.java
… fe/fe-core/src/main/java/com/starrocks/http/HttpConnectProcessor.java
… fe/fe-core/src/main/java/com/starrocks/qe/ConnectContext.java
… fe/fe-core/src/main/java/com/starrocks/qe/ConnectProcessor.java
… fe/fe-core/src/main/java/com/starrocks/qe/DefaultCoordinator.java
… fe/fe-core/src/main/java/com/starrocks/qe/HDFSBackendSelector.java
… fe/fe-core/src/main/java/com/starrocks/qe/SessionVariable.java
… fe/fe-core/src/main/java/com/starrocks/qe/StmtExecutor.java
… fe/fe-core/src/main/java/com/starrocks/scheduler/PartitionBasedMvRefreshProcessor.java
… fe/fe-core/src/main/java/com/starrocks/scheduler/SqlTaskRunProcessor.java
… fe/fe-core/src/main/java/com/starrocks/sql/InsertPlanner.java
… fe/fe-core/src/main/java/com/starrocks/sql/PlannerProfile.java
… fe/fe-core/src/main/java/com/starrocks/sql/StatementPlanner.java
… fe/fe-core/src/main/java/com/starrocks/sql/ast/StatementBase.java
… fe/fe-core/src/main/java/com/starrocks/sql/common/DebugOperatorTracer.java
… fe/fe-core/src/main/java/com/starrocks/sql/common/DebugRelationTracer.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/Group.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/GroupExpression.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/MvRewritePreprocessor.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/OptExpression.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/Optimizer.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/OptimizerContext.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/OptimizerTraceInfo.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/OptimizerTraceUtil.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rewrite/OptExternalPartitionPruner.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/implementation/LimitImplementationRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/join/ReorderJoinRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/SplitScanORToUnionRule.java