Fixes #issue

Support more `trace` command for debug:

core file: the new file on `profile` dir, the interface is `Tracer`, `Tracers`
other files just replace to the new code

* Refactor `PlannerProfile` and `OptimizerTraceUtils` to a common util
* Remove some duplicate implementations 
* Remove session variables `enableOptimizerTraceLog`,`enableMVOptimizerTraceLog`
* Limit Group/GroupExpression debug log output, default 3 line

other: 
* support trace by `LOGS`,`ALL`,`VALUES`,`TIMES` and different module
* dynamic trace query accord to the command, and not affect any performance by default 

we can use `Tracer` or `Tracers` in anywhere to debug code

examples, trace base times
```
MySQL td> trace times select * from t1;
+--------------------------------------------------+
| Explain String                                   |
+--------------------------------------------------+
| 0ms|-- Total[1] 2ms                              |
| 0ms|    -- Analyzer[1] 0                         |
| 1ms|    -- Transformer[1] 0                      |
| 1ms|    -- Optimizer[1] 1ms                      |
| 1ms|        -- preprocessMvs[1] 0                |
| 1ms|        -- RuleBaseOptimize[1] 0             |
| 2ms|        -- CostBaseOptimize[1] 0             |
| 2ms|        -- PhysicalRewrite[1] 0              |
| 2ms|        -- PlanValidate[1] 0                 |
| 2ms|            -- InputDependenciesChecker[1] 0 |
| 2ms|            -- TypeChecker[1] 0              |
| 2ms|            -- CTEUniqueChecker[1] 0         |
| 2ms|    -- ExecPlanBuild[1] 0                    |
| Trace Timer Cost: 23us                           |
+--------------------------------------------------+
14 rows in set
Time: 0.011s
```

trace detail module times
```
MySQL td> trace times optimizer select * from t1;
+--------------------------------------------------------+
| Explain String                                         |
+--------------------------------------------------------+
| 0ms|-- Total[1] 2ms                                    |
| 0ms|    -- Analyzer[1] 0                               |
| 0ms|    -- Transformer[1] 0                            |
| 1ms|    -- Optimizer[1] 1ms                            |
| 1ms|        -- preprocessMvs[1] 0                      |
| 1ms|        -- RuleBaseOptimize[1] 0                   |
| 1ms|            -- RewriteTreeTask[40] 0               |
| 1ms|                -- MergeTwoProjectRule[1] 0        |
| 1ms|                -- PruneProjectColumnsRule[2] 0    |
.............
| 2ms|        -- CostBaseOptimize[1] 0                   |
| 2ms|            -- OptimizeGroupTask[1] 0              |
| 2ms|            -- OptimizeExpressionTask[1] 0         |
| 2ms|            -- DeriveStatsTask[1] 0                |
| 2ms|            -- ApplyRuleTask[1] 0                  |
| 2ms|                -- OlapScanImplementationRule[1] 0 |
| 2ms|            -- EnforceAndCostTask[1] 0             |
| 2ms|        -- PhysicalRewrite[1] 0                    |
| 2ms|        -- PlanValidate[1] 0                       |
| 2ms|            -- InputDependenciesChecker[1] 0       |
| 2ms|            -- TypeChecker[1] 0                    |
| 2ms|            -- CTEUniqueChecker[1] 0               |
| 2ms|    -- ExecPlanBuild[1] 0                          |
| Trace Timer Cost: 53us                                 |
+--------------------------------------------------------+
29 rows in set
Time: 0.011s
MySQL td> trace times mv select t1.v2, sum(t1.v3) from t1 join t0 on t1.v1 = t0.v1 group by t1.v2
+--------------------------------------------------+
| Explain String                                   |
+--------------------------------------------------+
| 1ms|-- Total[1] 6ms                              |
| 1ms|    -- Analyzer[1] 0                         |
| 2ms|    -- Transformer[1] 0                      |
| 2ms|    -- Optimizer[1] 5ms                      |
| 2ms|        -- preprocessMvs[1] 0                |
| 3ms|        -- RuleBaseOptimize[1] 1ms           |
| 4ms|        -- CostBaseOptimize[1] 3ms           |
| 8ms|        -- PhysicalRewrite[1] 0              |
| 8ms|        -- PlanValidate[1] 0                 |
| 8ms|            -- InputDependenciesChecker[1] 0 |
| 8ms|            -- TypeChecker[1] 0              |
| 8ms|            -- CTEUniqueChecker[1] 0         |
| 8ms|    -- ExecPlanBuild[1] 0                    |
| Trace Timer Cost: 19us                           |
+--------------------------------------------------+
14 rows in set
Time: 0.016s
```

trace log in command
```
MySQL td> set trace_log_mode='command'
Query OK, 0 rows affected
Time: 0.002s
MySQL td> trace logs optimizer select * from t1;
+-----------------------------------------------------------------------------------------------------------------------------------------------------------+
| Explain String                                                                                                                                            |
+-----------------------------------------------------------------------------------------------------------------------------------------------------------+
| 1ms|    origin logicOperatorTree:                                                                                                                         |
| LogicalProjectOperator {projection=[1: v1, 2: v2, 3: v3]}                                                                                                 |
| ->  LogicalProjectOperator {projection=[1: v1, 2: v2, 3: v3]}                                                                                             |
|     ->  LogicalOlapScanOperator {table=7081712, selectedPartitionId=null, outputColumns=[1: v1, 2: v2, 3: v3], predicate=null, limit=-1}                  |
| 2ms|    [TRACE QUERY 0b5a9870-4701-11ee-9146-2e95bcb5f199] APPLY RULE TF_MERGE_TWO_PROJECT 72                                                             |
| Original Expression:                                                                                                                                      |
| LogicalProjectOperator {projection=[1: v1, 2: v2, 3: v3]}                                                                                                 |
| ->  LogicalProjectOperator {projection=[1: v1, 2: v2, 3: v3]}                                                                                             |
|     ->  LogicalOlapScanOperator {table=7081712, selectedPartitionId=null, outputColumns=[1: v1, 2: v2, 3: v3], predicate=null, limit=-1}                  |
| New Expression:                                                                                                                                           |
| 0:LogicalProjectOperator {projection=[1: v1, 2: v2, 3: v3]}                                                                                               |
| ->  LogicalOlapScanOperator {table=7081712, selectedPartitionId=null, outputColumns=[1: v1, 2: v2, 3: v3], predicate=null, limit=-1}                      |
|  ....................  |
| 3ms|    [TRACE QUERY 0b5a9870-4701-11ee-9146-2e95bcb5f199] APPLY RULE IMP_OLAP_LSCAN_TO_PSCAN 123                                                         |
| Original Expression:                                                                                                                                      |
| LogicalOlapScanOperator {table=7081712, selectedPartitionId=[7081711], outputColumns=[1: v1, 2: v2, 3: v3], predicate=null, limit=-1}                     |
| New Expression:                                                                                                                                           |
| 0:PhysicalOlapScanOperator {table=7081712, selectedPartitionId=[7081711], outputColumns=[1: v1, 2: v2, 3: v3], projection=null, predicate=null, limit=-1} |
| 4ms|    after extract best plan:                                                                                                                          |
| PhysicalOlapScanOperator {table=7081712, selectedPartitionId=[7081711], outputColumns=[1: v1, 2: v2, 3: v3], projection=null, predicate=null, limit=-1}   |
| 4ms|    final plan after physical rewrite:                                                                                                                |
| PhysicalOlapScanOperator {table=7081712, selectedPartitionId=[7081711], outputColumns=[1: v1, 2: v2, 3: v3], projection=null, predicate=null, limit=-1}   |
| Trace Timer Cost: 511us                                                                                                                                   |
+-----------------------------------------------------------------------------------------------------------------------------------------------------------+
89 rows in set
Time: 0.017s
MySQL td>
```

or trace logs to fe.log
```
MySQL td> set trace_log_mode='file'
Query OK, 0 rows affected
Time: 0.002s
MySQL td> trace logs optimizer select * from t1;
+--------------------------+
| Explain String           |
+--------------------------+
| Trace Timer Cost: 1858us |
+--------------------------+
1 row in set
Time: 0.012s
```
![image](https://github.com/StarRocks/starrocks/assets/5609319/f3177dd0-c521-4961-b20b-e9c5dfbe1e74)

trace different module:
```
MySQL td> CREATE MATERIALIZED VIEW mv1 as select v2, sum(v3) from t1 group by v2;
Query OK, 0 rows affected
Time: 0.024s
MySQL td> trace logs mv select v2, sum(v3) from t1 group by v2
+------------------------------------------------------------------------------------------------------------------------------------+
| Explain String                                                                                                                     |
+------------------------------------------------------------------------------------------------------------------------------------+
| 3ms|    [MV TRACE] [PREPARE 635622e0-4702-11ee-9146-2e95bcb5f199][GLOBAL] [SYNC=false] There are no related mvs for the query plan |
| 5ms|    [MV TRACE] [PREPARE 635622e0-4702-11ee-9146-2e95bcb5f199][GLOBAL] [SYNC=true] There are no related mvs for the query plan  |
| 8ms|    Query cannot be rewritten, please check the trace logs or `set enable_mv_optimizer_trace_log=on` to find more infos.       |
| Trace Timer Cost: 2073us                                                                                                           |
+------------------------------------------------------------------------------------------------------------------------------------+
4 rows in set
Time: 0.015s
```

trace the record values
```
MySQL td> CREATE MATERIALIZED VIEW mv2 DISTRIBUTED BY HASH(`v2`)
REFRESH ASYNC  as select t1.v2, sum(t1.v3) from t1 join t0 on t1.v1 = t0.v1 group by t1.v2 ;
Query OK, 0 rows affected
Time: 0.042s
MySQL td> trace values mv select t1.v2, sum(t1.v3) from t1 join t0 on t1.v1 = t0.v1 group by t1.v2
+----------------------------+
| Explain String             |
+----------------------------+
|  9ms| mv2: Rewrite Succeed |
| Trace Timer Cost: 24us     |
+----------------------------+
2 rows in set
Time: 0.018s
```

and in profile
```
MySQL emr_test.ssb_100g_orc_zlib> select count(1) from lineorder_flat;
+-----------+
| count(1)  |
+-----------+
| 600037902 |
+-----------+
1 row in set
Time: 4.712s
```
![image](https://github.com/StarRocks/starrocks/assets/5609319/cf642c76-8fda-4e80-bdcb-4d0cf310fa15)


## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [ ] Enhancement
- [x] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:

- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4