@wanpengfei-git 2023-08-30
#### [FE Incremental Coverage Report]

:disappointed: **fail** : 406 / 684 (59.36%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/sql/optimizer/MvRewritePreprocessor.java|0|1|00.00%|[247]|
|:large_blue_circle:|com/starrocks/sql/optimizer/GroupExpression.java|0|2|00.00%|[318, 322]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/implementation/LimitImplementationRule.java|0|1|00.00%|[41]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/materialization/rule/SingleTableMvRewriteRule.java|0|1|00.00%|[65]|
|:large_blue_circle:|com/starrocks/sql/common/DebugRelationTracer.java|0|27|00.00%|[27, 30, 35, 40, 41, 48, 50, 52, 54, 56, 58, 60, 66, 67, 68, 69, 75, 76, 78, 84, 85, 86, 87, 88, 89, 90, 95]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/SplitScanORToUnionRule.java|0|1|00.00%|[85]|
|:large_blue_circle:|com/starrocks/sql/common/DebugOperatorTracer.java|28|151|18.54%|[96, 97, 98, 104, 119, 120, 121, 122, 128, 129, 130, 131, 132, 133, 138, 139, 140, 141, 147, 148, 149, 150, 156, 161, 162, 163, 164, 170, 171, 172, 173, 192, 193, 194, 199, 200, 201, 202, 208, 213, 214, 215, 216, 217, 218, 219, 224, 225, 230, 231, 236, 237, 242, 247, 257, 262, 263, 269, 274, 279, 284, 296, 297, 298, 299, 318, 334, 335, 336, 337, 343, 344, 345, 346, 347, 348, 353, 354, 355, 356, 362, 367, 368, 369, 370, 376, 377, 378, 379, 385, 390, 391, 392, 393, 399, 400, 401, 402, 408, 413, 418, 419, 424, 425, 430, 431, 437, 438, 439, 440, 442, 443, 444, 446, 447, 452, 457, 462, 467, 479, 484, 489, 494]|
|:large_blue_circle:|com/starrocks/connector/iceberg/IcebergMetadata.java|1|5|20.00%|[267, 268, 269, 270]|
|:large_blue_circle:|com/starrocks/connector/hive/HiveMetaClient.java|5|20|25.00%|[121, 127, 140, 155, 163, 170, 177, 197, 204, 211, 227, 228, 261, 273, 275]|
|:large_blue_circle:|com/starrocks/common/profile/Timer.java|2|7|28.57%|[25, 29, 35, 39, 43]|
|:large_blue_circle:|com/starrocks/sql/optimizer/validate/InputDependenciesChecker.java|1|3|33.33%|[181, 189]|
|:large_blue_circle:|com/starrocks/sql/optimizer/Group.java|1|3|33.33%|[378, 381]|
|:large_blue_circle:|com/starrocks/qe/StmtExecutor.java|5|11|45.45%|[308, 561, 1413, 1415, 1417, 1419]|
|:large_blue_circle:|com/starrocks/scheduler/PartitionBasedMvRefreshProcessor.java|3|6|50.00%|[247, 248, 251]|
|:large_blue_circle:|com/starrocks/common/profile/VarTracer.java|11|22|50.00%|[35, 36, 37, 39, 40, 41, 42, 61, 62, 66, 67]|
|:large_blue_circle:|com/starrocks/sql/optimizer/OptimizerTraceInfo.java|1|2|50.00%|[41]|
|:large_blue_circle:|com/starrocks/common/profile/LogTracer.java|7|11|63.64%|[23, 26, 29, 32]|
|:large_blue_circle:|com/starrocks/common/profile/TracerImpl.java|82|126|65.08%|[79, 80, 81, 82, 114, 115, 117, 118, 166, 167, 169, 170, 171, 172, 173, 174, 175, 176, 177, 178, 179, 181, 183, 187, 188, 189, 190, 192, 196, 197, 198, 199, 202, 203, 204, 205, 206, 207, 208, 209, 210, 213, 214, 215]|
|:large_blue_circle:|com/starrocks/sql/optimizer/validate/MVRewriteValidator.java|10|15|66.67%|[49, 50, 69, 70, 82]|
|:large_blue_circle:|com/starrocks/common/profile/Tracer.java|10|15|66.67%|[27, 58, 62, 66, 70]|
|:large_blue_circle:|com/starrocks/sql/optimizer/OptimizerTraceUtil.java|18|24|75.00%|[34, 38, 69, 70, 71, 72]|
|:large_blue_circle:|com/starrocks/common/profile/Var.java|9|11|81.82%|[30, 46]|
|:large_blue_circle:|com/starrocks/common/profile/Tracers.java|81|88|92.05%|[85, 140, 141, 142, 180, 181, 182]|
|:large_blue_circle:|com/starrocks/sql/optimizer/Optimizer.java|16|16|100.00%|[]|
|:large_blue_circle:|com/starrocks/scheduler/SqlTaskRunProcessor.java|2|2|100.00%|[]|
|:large_blue_circle:|com/starrocks/qe/SessionVariable.java|4|4|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/RewriteTreeTask.java|3|3|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/ApplyRuleTask.java|4|4|100.00%|[]|
|:large_blue_circle:|com/starrocks/connector/RemoteFileOperations.java|2|2|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/materialization/MaterializedViewRewriter.java|2|2|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/OptExpression.java|8|8|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/validate/PlanValidator.java|2|2|100.00%|[]|
|:large_blue_circle:|com/starrocks/common/profile/TimeWatcher.java|35|35|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rewrite/OptExternalPartitionPruner.java|1|1|100.00%|[]|
|:large_blue_circle:|com/starrocks/common/profile/CommandLogTracer.java|11|11|100.00%|[]|
|:large_blue_circle:|com/starrocks/common/profile/FileLogTracer.java|10|10|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/SeriallyTaskScheduler.java|1|1|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/StatementPlanner.java|9|9|100.00%|[]|
|:large_blue_circle:|com/starrocks/http/HttpConnectProcessor.java|2|2|100.00%|[]|
|:large_blue_circle:|com/starrocks/qe/ConnectProcessor.java|3|3|100.00%|[]|
|:large_blue_circle:|com/starrocks/qe/DefaultCoordinator.java|3|3|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/join/ReorderJoinRule.java|1|1|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/ast/StatementBase.java|6|6|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/InsertPlanner.java|4|4|100.00%|[]|
|:large_blue_circle:|com/starrocks/qe/HDFSBackendSelector.java|2|2|100.00%|[]|